put winding count methods into category.
